requirements.txt
torch
torchvision
torchaudio
torchsummary
parselmouth
librosa
torchsummary
scikit-image
pydub
simpleaudio
nnaudio
numpy
imageio
pysox
pyaudio
